Refactoring: class GeoDataPointmap::InterpolateLineAddPointCommand moved to independent files.
